fix(themes): make the header files of the appropriate directory independent
Previously, the theme directory header files were
dependent on prism.hpp, which made it impossible
to include them without importing prism.hpp first.
This defeated the purpose of the header files. This
commit also fixes bugs in the IDE that stated
structures in the theme files could not be found.